AIOT_V1.1更新了mqtt服务器，待解决float不准问题

+++ include/sensor/include/sensor.h
@@ -6,11 +6,11 @@
 #include "json.h"
 
 typedef struct {
-    uint8_t dev_id;          // 设备号（0x01-0x03）
-    float temperature;       // 温度（°C）
-    float humidity;          // 湿度（%RH）
-    float co_ppm;            // CO浓度（ppm）
-    float light_lux;         // 光照度（lux）
+    int dev_id;          // 设备号（0x01-0x03）
+    int temperature;       // 温度（°C）
+    int humidity;          // 湿度（%RH）
+    int co_ppm;            // CO浓度（ppm）
+    int light_lux;         // 光照度（lux）
     pthread_mutex_t mutex;   // 数据访问互斥锁
 } SensorData;
